requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-htmx>=1.17.0
anthropic>=0.18.0
python-dotenv>=1.0.0
markdown>=3.5.0
bleach>=6.1.0
Pillow>=10.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
